refactor(activemodel,scripts): move HelperMethods to validations/helper-methods.ts
helper_methods.rb kebab-cases to helper-methods.ts (docs/ruby-ts-conventions.md),
which is where the fan-out-model-validates-of-macros-to-helper-methods story puts
_merge_attributes. The RUBY_FILE_TS_OVERRIDES entry follows the body, and
helper_methods.rb now maps 1:1 to a TS file instead of folding into validations.ts.
