+++ packages/activemodel/src/api.ts
@@ -20,8 +20,8 @@ import {
   contextForValidation as validationsContextForValidation,
   runValidationsBang as validationsRunValidationsBang,
   raiseValidationError as validationsRaiseValidationError,
-  HelperMethods as ValidationsHelperMethods,
 } from "./validations.js";
+import { HelperMethods } from "./validations/helper-methods.js";
 import {
   _assignAttributes as attrAssign,
   _assignAttribute as attrAssignOne,
@@ -56,7 +56,7 @@ export const raiseValidationError = validationsRaiseValidationError;
 /**
  * @internal Rails-private helper.
  */
-export const _mergeAttributes = ValidationsHelperMethods._mergeAttributes;
+export const _mergeAttributes = HelperMethods._mergeAttributes;
 
 /**
  * @internal Rails-private helper.
