+++ packages/activemodel/src/validations/exclusion.ts
@@ -1,7 +1,7 @@
 import { EachValidator } from "../validator.js";
 import type { ValidatableRecord } from "../validator.js";
 import { include } from "@blazetrails/activesupport";
-import type { AttrNameArg, HelperMethodsHost } from "../validations.js";
+import type { AttrNameArg, HelperMethodsHost } from "./helper-methods.js";
 import {
   checkValidityBang,
   type Clusivity,
